# openapi3.yaml
openapi: 3.0.1
info:
  title: service-name
  description: basic template for map colonies service
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /anotherResource:
    get:
      operationId: getAnotherResource
      tags:
        - anotherResource
      summary: gets the resource
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/anotherResource'
  /resourceName:
    get:
      operationId: getResourceName
      tags:
        - resourceName
      summary: gets the resource
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resource'
    post:
      operationId: createResource
      tags:
        - resourceName
      summary: creates a new record of type resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/resource'
      responses:
        201:
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resource'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    resource:
      type: object
      required:
        - id
        - name
        - description
      properties:
        id:
          type: number
          format: int64
        name:
          type: string
        description:
          type: string
    anotherResource:
      type: object
      required:
        - kind
        - isAlive
      properties:
        kind:
          type: string
        isAlive:
          type: boolean